Version 1.0.8: Improved guest checkout and compatibility fixes
- Enhanced WooCommerce session initialization with proper cookie handling
- Better error messages showing actual WooCommerce errors
- Fixed WordPress 6.7+ translation loading warning
- Fixed PHP 8.1+ float to int deprecation in darken_color()

+++ includes/class-bulk-payment-admin.php
@@ -697,10 +697,10 @@ class Bulk_Payment_Admin {
             $b = hexdec(substr($hex, 4, 2));
         }
 
-        // Darken
-        $r = max(0, min(255, $r - ($r * $percent / 100)));
-        $g = max(0, min(255, $g - ($g * $percent / 100)));
-        $b = max(0, min(255, $b - ($b * $percent / 100)));
+        // Darken and cast to int to avoid PHP 8.1+ deprecation warnings
+        $r = (int) max(0, min(255, $r - ($r * $percent / 100)));
+        $g = (int) max(0, min(255, $g - ($g * $percent / 100)));
+        $b = (int) max(0, min(255, $b - ($b * $percent / 100)));
 
         // Convert back to hex
         $r = str_pad(dechex($r), 2, '0', STR_PAD_LEFT);
